+++ ckan/logic/auth/get.py
@@ -59,8 +59,8 @@ def organization_list(context, data_dict):
 def organization_list_for_user(context, data_dict):
     return {'success': True}
 
-def licence_list(context, data_dict):
-    # Licences list is visible by default
+def license_list(context, data_dict):
+    # Licenses list is visible by default
     return {'success': True}
 
 def tag_list(context, data_dict):
@@ -100,6 +100,8 @@ def package_show(context, data_dict):
         auth = new_authz.is_authorized('package_update',
                                        context, data_dict)
         authorized = auth.get('success')
+    elif package.owner_org is None and package.state == 'active':
+        return {'success': True}
     else:
         # anyone can see a public package
         if not package.private and package.state == 'active':
@@ -133,7 +135,7 @@ def resource_show(context, data_dict):
     authorized = package_show(context, pkg_dict).get('success')
 
     if not authorized:
-        return {'success': False, 'msg': _('User %s not authorized to read resource %s') % (str(user), resource.id)}
+        return {'success': False, 'msg': _('User %s not authorized to read resource %s') % (user, resource.id)}
     else:
         return {'success': True}
 
@@ -191,10 +193,10 @@ def tag_show_rest(context, data_dict):
     return tag_show(context, data_dict)
 
 def get_site_user(context, data_dict):
-    if not context.get('ignore_auth'):
-        return {'success': False, 'msg': 'Only internal services allowed to use this action'}
-    else:
-        return {'success': True}
+    # FIXME this is available to sysadmins currently till
+    # @auth_sysadmins_check decorator is added
+    return {'success': False,
+            'msg': 'Only internal services allowed to use this action'}
 
 
 def member_roles_list(context, data_dict):
@@ -262,3 +264,11 @@ def dataset_followee_list(context, data_dict):
 
 def group_followee_list(context, data_dict):
     return _followee_list(context, data_dict)
+
+
+def user_reset(context, data_dict):
+    return {'success': True}
+
+
+def request_reset(context, data_dict):
+    return {'success': True}