+++ ckan/tests/functional/test_user.py
@@ -99,9 +99,10 @@ class TestUserController(FunctionalTestCase, HtmlCheckMethods, PylonsTestCase, S
         assert 'Don\'t have an OpenID' in res, res
 
     def test_logout(self):
-        res = self.app.get('/user/logout')
+        res = self.app.get('/user/_logout')
         res2 = res.follow()
-        res2 = res2.follow()
+        while res2.status == 302:
+            res2 = res2.follow()
         assert 'You have logged out successfully.' in res2, res2
 
     def _get_cookie_headers(self, res):
@@ -141,6 +142,7 @@ class TestUserController(FunctionalTestCase, HtmlCheckMethods, PylonsTestCase, S
                res.header('Location').startswith('/user/logged_in')
 
         # then get redirected to user's dashboard
+        res = res.follow()
         res = res.follow()
         assert_equal(res.status, 302)
         assert res.header('Location').startswith('http://localhost/dashboard') or \
@@ -213,6 +215,7 @@ class TestUserController(FunctionalTestCase, HtmlCheckMethods, PylonsTestCase, S
                res.header('Location').startswith('/user/logged_in')
 
         # then get redirected to login
+        res = res.follow()
         res = res.follow()
         assert_equal(res.status, 302)
         assert res.header('Location').startswith('http://localhost/user/login') or \
@@ -336,9 +339,10 @@ class TestUserController(FunctionalTestCase, HtmlCheckMethods, PylonsTestCase, S
         assert 'logout' in res.body, res.body
 
         # logout and login as user B
-        res = self.app.get('/user/logout')
+        res = self.app.get('/user/_logout')
         res2 = res.follow()
-        res2 = res2.follow()
+        while res2.status == 302:
+            res2 = res2.follow()
         assert 'You have logged out successfully.' in res2, res2
         offset = url_for(controller='user', action='login')
         res = self.app.get(offset)
@@ -776,8 +780,9 @@ class TestUserController(FunctionalTestCase, HtmlCheckMethods, PylonsTestCase, S
 
     def test_user_edit_unknown_user(self):
         offset = url_for(controller='user', action='edit', id='unknown_person')
-        res = self.app.get(offset, status=404)
-        assert 'User not found' in res, res
+        res = self.app.get(offset, status=302) # redirect to login page
+        res = res.follow()
+        assert 'Login' in res, res
 
     def test_user_edit_not_logged_in(self):
         # create user
@@ -854,16 +859,9 @@ class TestUserController(FunctionalTestCase, HtmlCheckMethods, PylonsTestCase, S
         assert 'No such user: unknown' in res, res # error
 
     def test_request_reset_user_password_using_search(self):
-        CreateTestData.create_user(name='larry1', email='kittens@john.com')
         offset = url_for(controller='user',
                          action='request_reset')
-        res = self.app.get(offset)
-        fv = res.forms['user-password-reset']
-        fv['user'] = 'kittens'
-        res = fv.submit()
-        assert_equal(res.status, 302)
-        assert_equal(res.header_dict['Location'], 'http://localhost/?__no_cache__=True')
-
+        CreateTestData.create_user(name='larry1', fullname='kittens')
         CreateTestData.create_user(name='larry2', fullname='kittens')
         res = self.app.get(offset)
         fv = res.forms['user-password-reset']