+++ doc/conf.py
@@ -50,15 +50,23 @@ rst_epilog = '''
 .. |reload_apache| replace:: sudo service apache2 reload
 .. |solr| replace:: Solr
 .. |restructuredtext| replace:: reStructuredText
+.. |nginx| replace:: Nginx
+.. |sqlite| replace:: SQLite
+.. |python| replace:: Python
+.. |sqlalchemy| replace:: SQLAlchemy
+.. |javascript| replace:: JavaScript
+.. |apache| replace:: Apache
 
 '''
 
 # Add any Sphinx extension module names here, as strings. They can be extensions
 # coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
 extensions = ['sphinx.ext.autodoc', 'sphinx.ext.todo']
-autodoc_member_order = 'bysource'
 
 autodoc_member_order = 'bysource'
+
+todo_include_todos = True
+
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']