+++ CHANGELOG.rst
@@ -1,8 +1,36 @@
-.. _changelog:
+.. This tocdepth stops Sphinx from putting every subsection title in this file
+   into the master table of contents.
+
+:tocdepth: 1
 
 ---------
 Changelog
 ---------
+
+v2.2
+====
+
+API changes and deprecations:
+
+* The edit() and after_update() methods of IPackageController plugins are now
+  called when updating a resource using the web frontend or the
+  resource_update API action [#1052]
+
+v2.0.1 2013-06-11
+=================
+
+Bug fixes:
+ * Use IDatasetForm schema for resource_update (#897)
+ * Fixes for CKAN being run on a non-root URL (#948, #913)
+ * Fix resource edit errors losing info (#580)
+ * Fix Czech translation (#900)
+ * Allow JSON filters for datastore_search on GET requests (#917)
+ * Install vdm from the Python Package Index (#764)
+ * Allow extra parameters on Solr queries (#739)
+ * Create site user at startup if it does not exist (#952)
+ * Fix modal popups positioning (#828)
+ * Fix wrong redirect on dataset form on IE (#963)
+
 
 v2.0 2013-05-10
 ===============
@@ -404,7 +432,7 @@ v1.5 2011-11-07
 Major:
  * New visual theme (#1108)
     * Package & Resource edit overhaul (#1294/#1348/#1351/#1368/#1296)
-    * JS and CSS reorganisation (#1282, #1349, #1380)
+    * JS and CSS reorganization (#1282, #1349, #1380)
  * Apache Solr used for search in core instead of Postgres (#1275, #1361, #1365)
  * Authorization system now embedded in the logic layer (#1253)
  * Captcha added for user registration (#1307, #1431)